Moving emissions into the emissions module
