+++ examples/parametric/plot_generate_lines.py
@@ -8,7 +8,7 @@ parametric galaxy.
 
 from unyt import Msun, Myr
 
-import synthesizer.line_ratios as line_ratios
+import synthesizer.emissions_ratios as line_ratios
 from synthesizer.emission_models import AttenuatedEmission, NebularEmission
 from synthesizer.emission_models.attenuation import PowerLaw
 from synthesizer.grid import Grid
