+++ src/synthesizer/emission_models/dust/emission.py
@@ -43,8 +43,8 @@ from unyt import (
 )
 
 from synthesizer import exceptions
+from synthesizer.emissions import Sed
 from synthesizer.grid import Grid
-from synthesizer.sed import Sed
 from synthesizer.synth_warnings import warn
 from synthesizer.units import accepts
 from synthesizer.utils import planck
